core/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# 科学计算
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# 数据处理
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 神经科学相关
nixio>=1.5.0b3
pyefd>=1.3.2

# 配置管理
pyyaml>=6.0
configparser>=5.3.0
click>=8.1.0

# 日志和监控
wandb>=0.15.0
tensorboard>=2.14.0
rich>=13.0.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# 可视化和交互
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0

# 性能优化
numba>=0.57.0
cython>=0.29.0
numba-cuda>=0.52.0

# 内存和性能分析
memory-profiler>=0.60.0
psutil>=5.9.0
py-spy>=0.3.14

# 分布式训练
deepspeed>=0.12.0
fairscale>=1.4.0

# 数据存储
h5py>=3.9.0
lmdb>=1.4.0
pickle5>=0.0.12

# 进度条
tqdm>=4.66.0
alive-progress>=3.0.1

# 机器学习工具
optuna>=3.3.0
ray>=2.6.0

# 图像处理
albumentations>=1.3.0
imageio>=2.31.0

# 生物信息学相关
networkx>=3.1
igraph>=0.10.0